+++ daemon/daemonapi/post_node_status_subscription.go
@@ -17,14 +17,14 @@ import (
 func (a *DaemonApi) PostNodeStatusSubscription(w http.ResponseWriter, r *http.Request, params PostNodeStatusSubscriptionParams) {
 	var (
 		handlerName = "PostNodeStatusFrozenSubscription"
-		limit       int64
+		limit       uint64
 		maxDuration = 5 * time.Second
 	)
 	log := getLogger(r, handlerName)
 	log.Debug().Msg("starting")
 	defer log.Debug().Msg("done")
 	if params.Limit != nil {
-		limit = *params.Limit
+		limit = uint64(*params.Limit)
 	}
 	if params.Duration != nil {
 		if v, err := converters.Duration.Convert(*params.Duration); err != nil {
